search is working - if you use a autocomlete entry and hit enter . Refactor MapComponent and LandingPage for improved map functionality
- Updated libraries in App.tsx to include 'marker' for enhanced map features.
- Removed unused Google Maps API loader from MapComponent and implemented marker management using useRef.
- Simplified LandingPage by replacing custom search input with a new SearchBar component for better location selection.
- Enhanced location update logic in MapComponent to set zoom level and manage marker lifecycle effectively.
