+++ src/components/map/MapComponent.tsx
@@ -1,5 +1,5 @@
-import { useCallback, useState, useEffect } from 'react';
-import { GoogleMap, useJsApiLoader, Marker } from '@react-google-maps/api';
+import { useCallback, useState, useEffect, useRef } from 'react';
+import { GoogleMap } from '@react-google-maps/api';
 
 const containerStyle = {
   width: '100%',
@@ -17,13 +17,8 @@ interface MapComponentProps {
 }
 
 const MapComponent = ({ onLoadCallback, selectedLocation }: MapComponentProps) => {
-  const { isLoaded, loadError } = useJsApiLoader({
-    id: 'google-map-script',
-    googleMapsApiKey: process.env.REACT_APP_GOOGLE_MAPS_API_KEY || '',
-    libraries: ['places']
-  });
-
   const [map, setMap] = useState<google.maps.Map | null>(null);
+  const markerRef = useRef<google.maps.marker.AdvancedMarkerElement | null>(null);
 
   const onLoad = useCallback((map: google.maps.Map) => {
     setMap(map);
@@ -33,18 +28,31 @@ const MapComponent = ({ onLoadCallback, selectedLocation }: MapComponentProps) =
   }, [onLoadCallback]);
 
   const onUnmount = useCallback(() => {
+    if (markerRef.current) {
+      markerRef.current.map = null;
+      markerRef.current = null;
+    }
     setMap(null);
   }, []);
 
   useEffect(() => {
     if (map && selectedLocation) {
-      console.log('MapComponent received location:', selectedLocation);
+      console.log('MapComponent updating location:', selectedLocation);
       map.panTo(selectedLocation);
+      map.setZoom(15);
+
+      // Remove existing marker
+      if (markerRef.current) {
+        markerRef.current.map = null;
+      }
+
+      // Create new marker
+      markerRef.current = new google.maps.marker.AdvancedMarkerElement({
+        map,
+        position: selectedLocation
+      });
     }
   }, [map, selectedLocation]);
-
-  if (loadError) return <div>Map cannot be loaded right now</div>;
-  if (!isLoaded) return <div>Loading...</div>;
 
   return (
     <GoogleMap
@@ -53,11 +61,10 @@ const MapComponent = ({ onLoadCallback, selectedLocation }: MapComponentProps) =
       zoom={13}
       onLoad={onLoad}
       onUnmount={onUnmount}
-    >
-      {selectedLocation && (
-        <Marker position={selectedLocation} />
-      )}
-    </GoogleMap>
+      options={{
+        mapId: process.env.REACT_APP_GOOGLE_MAPS_ID
+      }}
+    />
   );
 };
