+++ src/pages/LandingPage.tsx
@@ -1,104 +1,21 @@
-import { useState, useRef, useEffect } from 'react';
+import { useState } from 'react';
 import { 
   Box, 
   Button, 
   AppBar, 
   Toolbar, 
   Typography,
-  InputBase,
-  alpha,
-  styled
 } from '@mui/material';
-import SearchIcon from '@mui/icons-material/Search';
 import MapComponent from '../components/map/MapComponent';
 import AuthModal from '../components/auth/AuthModal';
 import { auth } from '../config/firebase';
 import { useAuthState } from 'react-firebase-hooks/auth';
-
-const Search = styled('div')(({ theme }) => ({
-  position: 'relative',
-  borderRadius: theme.shape.borderRadius,
-  backgroundColor: alpha(theme.palette.common.white, 0.15),
-  '&:hover': {
-    backgroundColor: alpha(theme.palette.common.white, 0.25),
-  },
-  marginRight: theme.spacing(2),
-  marginLeft: 0,
-  width: '100%',
-  [theme.breakpoints.up('sm')]: {
-    marginLeft: theme.spacing(3),
-    width: 'auto',
-  },
-}));
-
-const SearchIconWrapper = styled('div')(({ theme }) => ({
-  padding: theme.spacing(0, 2),
-  height: '100%',
-  position: 'absolute',
-  pointerEvents: 'none',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-}));
-
-const StyledInputBase = styled(InputBase)(({ theme }) => ({
-  color: 'inherit',
-  '& .MuiInputBase-input': {
-    padding: theme.spacing(1, 1, 1, 0),
-    paddingLeft: `calc(1em + ${theme.spacing(4)})`,
-    transition: theme.transitions.create('width'),
-    width: '100%',
-    [theme.breakpoints.up('md')]: {
-      width: '40ch',
-    },
-  },
-}));
+import SearchBar from '../components/search/SearchBar';
 
 const LandingPage = () => {
   const [authModalOpen, setAuthModalOpen] = useState(false);
   const [user] = useAuthState(auth);
-  const [searchInput, setSearchInput] = useState('');
   const [selectedLocation, setSelectedLocation] = useState<google.maps.LatLngLiteral | null>(null);
-  const searchInputRef = useRef<HTMLInputElement>(null);
-  const autocompleteRef = useRef<google.maps.places.Autocomplete | null>(null);
-
-  useEffect(() => {
-    if (searchInputRef.current && !autocompleteRef.current) {
-      autocompleteRef.current = new window.google.maps.places.Autocomplete(
-        searchInputRef.current,
-        {
-          types: ['establishment'],
-          componentRestrictions: { country: 'ca' },
-          fields: ['geometry', 'name', 'formatted_address']
-        }
-      );
-
-      autocompleteRef.current.addListener('place_changed', () => {
-        const place = autocompleteRef.current?.getPlace();
-        console.log('Selected place:', place);
-
-        if (place?.geometry?.location) {
-          const location = {
-            lat: place.geometry.location.lat(),
-            lng: place.geometry.location.lng()
-          };
-          console.log('Setting new location:', location);
-          setSelectedLocation(location);
-          setSearchInput(place.formatted_address || '');
-        }
-      });
-    }
-
-    return () => {
-      if (autocompleteRef.current) {
-        google.maps.event.clearInstanceListeners(autocompleteRef.current);
-      }
-    };
-  }, [searchInputRef.current]);
-
-  const handleSearchChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setSearchInput(event.target.value);
-  };
 
   return (
     <Box sx={{ height: '100vh', display: 'flex', flexDirection: 'column' }}>
@@ -108,35 +25,16 @@ const LandingPage = () => {
             Hockey Rink Map
           </Typography>
           
-          {user && (
-            <Search>
-              <SearchIconWrapper>
-                <SearchIcon />
-              </SearchIconWrapper>
-              <StyledInputBase
-                inputRef={searchInputRef}
-                value={searchInput}
-                onChange={handleSearchChange}
-                placeholder="Search for rinks…"
-                inputProps={{ 'aria-label': 'search' }}
-              />
-            </Search>
-          )}
+          {user && <SearchBar onLocationSelect={setSelectedLocation} />}
 
           <Box sx={{ flexGrow: 1 }} />
           
           {user ? (
-            <Button 
-              color="inherit" 
-              onClick={() => auth.signOut()}
-            >
+            <Button color="inherit" onClick={() => auth.signOut()}>
               Logout
             </Button>
           ) : (
-            <Button 
-              color="inherit" 
-              onClick={() => setAuthModalOpen(true)}
-            >
+            <Button color="inherit" onClick={() => setAuthModalOpen(true)}>
               Login / Sign Up
             </Button>
           )}
@@ -144,9 +42,7 @@ const LandingPage = () => {
       </AppBar>
       
       <Box sx={{ flexGrow: 1 }}>
-        <MapComponent 
-          selectedLocation={selectedLocation}
-        />
+        <MapComponent selectedLocation={selectedLocation} />
       </Box>
 
       <AuthModal
